fine-grained control of mouse sensitivity
A backwards-compatible change... mouse movements multiplied by 2^(sensitivity) instead of bit shifted.  For example, if you want a 52% increase in mouse sensitivity then set your mouseSensitivity to 0.6.

+++ src/game/CAvaraGame.h
@@ -158,7 +158,7 @@ public:
     Handle mapRes; //	Keyboard mapping resource handle.
 
     short moJoOptions; //	Mouse and Joystick options.
-    short sensitivity;
+    double sensitivity;
 
     double latencyTolerance;
